+++ prow/config/topology/external-istiod.json
@@ -1,0 +1,32 @@
+[
+  {
+    "kind": "Kubernetes",
+    "clusterName": "external",
+    "podSubnet": "10.10.0.0/16",
+    "svcSubnet": "10.255.10.0/24",
+    "network": "network-1",
+    "primaryClusterName": "external",
+    "configClusterName": "config"
+  },
+  {
+    "kind": "Kubernetes",
+    "clusterName": "config",
+    "podSubnet": "10.20.0.0/16",
+    "svcSubnet": "10.255.20.0/24",
+    "network": "network-1",
+    "primaryClusterName": "external",
+    "configClusterName": "config",
+    "meta": {
+      "fakeVM": false
+    }
+  },
+  {
+    "kind": "Kubernetes",
+    "clusterName": "additional",
+    "podSubnet": "10.30.0.0/16",
+    "svcSubnet": "10.255.30.0/24",
+    "network": "network-2",
+    "primaryClusterName": "external",
+    "configClusterName": "config"
+  }
+]
