+++ tests/integration/external/testdata/gateway.yaml
